新增自增,自减 和已知bug修复

+++ api/delete.js
@@ -2,10 +2,11 @@ const query = require('./query')
 
 module.exports = (mysql, pk) => {
     if (pk != undefined) {
-        let id = mysql.getPrimaryKey()
-        mysql.sql = 'DELETE FROM ' + '`' + mysql.tableName + '` WHERE `' + id + '` = \'' + pk + '\''
+        mysql.sql = 'DELETE FROM ' + '`' + mysql.tableName + '` WHERE `' + mysql.primaryKey + '` = \'' + pk + '\''
     } else {
         let wheres = mysql.getWheres()
+        if (wheres === '')
+            throw 'Lack of renewal conditions'
         mysql.sql = 'DELETE FROM ' + '`' + mysql.tableName + '`' + wheres
     }
     return query(mysql, mysql.sql);
